requirements.txt
networkx == 3.1
numpy == 1.25.2
pandas == 2.0.3
scipy == 1.9.3
cython == 3.0.2
cysignals == 1.11.2
matplotlib == 3.9.0
PyYAML == 6.0.1